fixed init middleware promise bug and tested validation with  user fields
