+++ lib/middleware/initializeMiddleware.js
@@ -1,6 +1,6 @@
 export default function initializeMiddleware(middleware) {
 	return (req, res) => {
-		new Promise((resolve, reject) => {
+		return new Promise((resolve, reject) => {
 			middleware(req, res, (result) => {
 				if (result instanceof Error) {
 					return reject(result);
